+++ R/verifiy_data.R
@@ -1,0 +1,18 @@
+# TODO: add tests.
+verify_iao_data <- function(dt) {
+    # TODO: perform verification with functions found below.
+    # TODO: concatenate the error message.
+    list("is_ok" = TRUE, "error_message" = "")
+}
+
+
+# TODO: add tests.
+verify_colnames <- function() {
+    # TODO: check if columns exist.
+}
+
+
+# TODO: add tests.
+verify_column_data <- function() {
+    # TODO: check column elements types.
+}